Refactor runs into modules
