+++ src/dstack/_internal/server/background/tasks/process_submitted_jobs.py
@@ -1,6 +1,5 @@
 import asyncio
 import itertools
-import math
 import uuid
 from contextlib import AsyncExitStack
 from datetime import datetime, timedelta
@@ -25,7 +24,6 @@ from dstack._internal.core.errors import BackendError, ServerClientError
 from dstack._internal.core.models.common import NetworkMode
 from dstack._internal.core.models.compute_groups import ComputeGroupProvisioningData
 from dstack._internal.core.models.fleets import (
-    Fleet,
     FleetConfiguration,
     FleetNodesSpec,
     FleetSpec,
@@ -74,17 +72,15 @@ from dstack._internal.server.models import (
 )
 from dstack._internal.server.services.backends import get_project_backend_by_type_or_error
 from dstack._internal.server.services.fleets import (
+    check_can_create_new_cloud_instance_in_fleet,
     fleet_model_to_fleet,
     generate_fleet_name,
-    get_fleet_requirements,
+    get_fleet_master_instance_provisioning_data,
     get_next_instance_num,
     is_cloud_cluster,
 )
 from dstack._internal.server.services.instances import (
-    filter_pool_instances,
-    get_instance_offer,
     get_instance_provisioning_data,
-    get_shared_pool_instances_with_offers,
 )
 from dstack._internal.server.services.jobs import (
     check_can_attach_job_volumes,
@@ -107,14 +103,16 @@ from dstack._internal.server.services.placement import (
     placement_group_model_to_placement_group_optional,
     schedule_fleet_placement_groups_deletion,
 )
-from dstack._internal.server.services.requirements.combine import (
-    combine_fleet_and_run_profiles,
-    combine_fleet_and_run_requirements,
-)
 from dstack._internal.server.services.runs import (
-    check_run_spec_requires_instance_mounts,
-    get_nodes_required_num_for_run,
     run_model_to_run,
+)
+from dstack._internal.server.services.runs.plan import (
+    find_optimal_fleet_with_offers,
+    get_run_profile_and_requirements_in_fleet,
+)
+from dstack._internal.server.services.runs.spec import (
+    check_run_spec_requires_instance_mounts,
+    get_nodes_required_num,
 )
 from dstack._internal.server.services.volumes import (
     volume_model_to_volume,
@@ -320,7 +318,7 @@ async def _process_submitted_job(
                 instance_filters=instance_filters,
             )
         fleet_models = fleet_models_with_instances + fleet_models_without_instances
-        fleet_model, fleet_instances_with_offers, _ = await _find_optimal_fleet_with_offers(
+        fleet_model, fleet_instances_with_offers, _ = await find_optimal_fleet_with_offers(
             project=project,
             fleet_models=fleet_models,
             run_model=run_model,
@@ -598,161 +596,6 @@ async def _refetch_fleet_models_with_instances(
     return fleet_models
 
 
-async def _find_optimal_fleet_with_offers(
-    project: ProjectModel,
-    fleet_models: list[FleetModel],
-    run_model: Optional[RunModel],
-    run_spec: RunSpec,
-    job: Job,
-    master_job_provisioning_data: Optional[JobProvisioningData],
-    volumes: Optional[list[list[Volume]]],
-) -> tuple[
-    Optional[FleetModel],
-    list[tuple[InstanceModel, InstanceOfferWithAvailability]],
-    list[tuple[Backend, InstanceOfferWithAvailability]],
-]:
-    """
-    Finds the optimal fleet for the run among the given fleet models and returns
-    the fleet model, pool offers with instances, and backend offers.
-    Returns empty backend offers if run_model.fleet is set since
-    backend offer from this function are needed only for run plan.
-    """
-    if run_model is not None and run_model.fleet is not None:
-        # Using the fleet that was already chosen by the master job
-        fleet_instance_offers = _get_fleet_instance_offers(
-            fleet_model=run_model.fleet,
-            run_spec=run_spec,
-            job=job,
-            master_job_provisioning_data=master_job_provisioning_data,
-            volumes=volumes,
-        )
-        return run_model.fleet, fleet_instance_offers, []
-
-    nodes_required_num = get_nodes_required_num_for_run(run_spec)
-    # The current strategy is first to consider fleets that can accommodate
-    # the run without additional provisioning and choose the one with the cheapest pool offer.
-    # Then choose a fleet with the cheapest pool offer among all fleets with pool offers.
-    # If there are no fleets with pool offers, choose a fleet with a cheapest backend offer.
-    # Fallback to autocreated fleet if fleets have no pool or backend offers.
-    # TODO: Consider trying all backend offers and then choosing a fleet.
-    candidate_fleets_with_offers: list[
-        tuple[
-            Optional[FleetModel],
-            list[tuple[InstanceModel, InstanceOfferWithAvailability]],
-            list[tuple[Backend, InstanceOfferWithAvailability]],
-            int,
-            int,
-            tuple[int, float, float],
-        ]
-    ] = []
-    for candidate_fleet_model in fleet_models:
-        candidate_fleet = fleet_model_to_fleet(candidate_fleet_model)
-        if (
-            is_multinode_job(job)
-            and candidate_fleet.spec.configuration.placement != InstanceGroupPlacement.CLUSTER
-        ):
-            # Limit multinode runs to cluster fleets to guarantee best connectivity.
-            continue
-
-        fleet_instance_offers = _get_fleet_instance_offers(
-            fleet_model=candidate_fleet_model,
-            run_spec=run_spec,
-            job=job,
-            # No need to pass master_job_provisioning_data for master job
-            # as all pool offers are suitable.
-            master_job_provisioning_data=None,
-            volumes=volumes,
-        )
-        fleet_has_pool_capacity = nodes_required_num <= len(fleet_instance_offers)
-        fleet_cheapest_instance_offer = math.inf
-        if len(fleet_instance_offers) > 0:
-            fleet_cheapest_instance_offer = fleet_instance_offers[0][1].price
-
-        try:
-            _check_can_create_new_instance_in_fleet(candidate_fleet)
-            profile, requirements = _get_run_profile_and_requirements_in_fleet(
-                job=job,
-                run_spec=run_spec,
-                fleet=candidate_fleet,
-            )
-        except ValueError:
-            fleet_backend_offers = []
-        else:
-            # Master job offers must be in the same cluster as existing instances.
-            master_instance_provisioning_data = _get_fleet_master_instance_provisioning_data(
-                fleet_model=candidate_fleet_model,
-                fleet_spec=candidate_fleet.spec,
-            )
-            # Handle multinode for old jobs that don't have requirements.multinode set.
-            # TODO: Drop multinode param.
-            multinode = requirements.multinode or is_multinode_job(job)
-            fleet_backend_offers = await get_offers_by_requirements(
-                project=project,
-                profile=profile,
-                requirements=requirements,
-                exclude_not_available=True,
-                multinode=multinode,
-                master_job_provisioning_data=master_instance_provisioning_data,
-                volumes=volumes,
-                privileged=job.job_spec.privileged,
-                instance_mounts=check_run_spec_requires_instance_mounts(run_spec),
-            )
-
-        fleet_cheapest_backend_offer = math.inf
-        if len(fleet_backend_offers) > 0:
-            fleet_cheapest_backend_offer = fleet_backend_offers[0][1].price
-
-        if not _run_can_fit_into_fleet(run_spec, candidate_fleet):
-            logger.debug("Skipping fleet %s from consideration: run cannot fit into fleet")
-            continue
-
-        fleet_priority = (
-            not fleet_has_pool_capacity,
-            fleet_cheapest_instance_offer,
-            fleet_cheapest_backend_offer,
-        )
-        candidate_fleets_with_offers.append(
-            (
-                candidate_fleet_model,
-                fleet_instance_offers,
-                fleet_backend_offers,
-                len(fleet_instance_offers),
-                len(fleet_backend_offers),
-                fleet_priority,
-            )
-        )
-    if len(candidate_fleets_with_offers) == 0:
-        return None, [], []
-    if (
-        not FeatureFlags.AUTOCREATED_FLEETS_DISABLED
-        and run_spec.merged_profile.fleets is None
-        and all(t[3] == 0 and t[4] == 0 for t in candidate_fleets_with_offers)
-    ):
-        # If fleets are not specified and no fleets have available pool
-        # or backend offers, create a new fleet.
-        # This is for compatibility with non-fleet-first UX when runs created new fleets
-        # if there are no instances to reuse.
-        return None, [], []
-    candidate_fleets_with_offers.sort(key=lambda t: t[-1])
-    return candidate_fleets_with_offers[0][:3]
-
-
-def _get_fleet_master_instance_provisioning_data(
-    fleet_model: FleetModel,
-    fleet_spec: FleetSpec,
-) -> Optional[JobProvisioningData]:
-    master_instance_provisioning_data = None
-    if fleet_spec.configuration.placement == InstanceGroupPlacement.CLUSTER:
-        # Offers for master jobs must be in the same cluster as existing instances.
-        fleet_instance_models = [im for im in fleet_model.instances if not im.deleted]
-        if len(fleet_instance_models) > 0:
-            master_instance_model = fleet_instance_models[0]
-            master_instance_provisioning_data = JobProvisioningData.__response__.parse_raw(
-                master_instance_model.job_provisioning_data
-            )
-    return master_instance_provisioning_data
-
-
 async def _fetch_fleet_with_master_instance_provisioning_data(
     exit_stack: AsyncExitStack,
     session: AsyncSession,
@@ -801,84 +644,11 @@ async def _fetch_fleet_with_master_instance_provisioning_data(
                     .execution_options(populate_existing=True)
                 )
                 fleet_model = res.unique().scalar_one()
-            master_instance_provisioning_data = _get_fleet_master_instance_provisioning_data(
+            master_instance_provisioning_data = get_fleet_master_instance_provisioning_data(
                 fleet_model=fleet_model,
                 fleet_spec=fleet.spec,
             )
     return master_instance_provisioning_data
-
-
-def _run_can_fit_into_fleet(run_spec: RunSpec, fleet: Fleet) -> bool:
-    """
-    Returns `False` if the run cannot fit into fleet for sure.
-    This is helpful heuristic to avoid even considering fleets too small for a run.
-    A run may not fit even if this function returns `True`.
-    This will lead to some jobs failing due to exceeding `nodes.max`
-    or more than `nodes.max` instances being provisioned
-    and eventually removed by the fleet consolidation logic.
-    """
-    # No check for cloud fleets with blocks > 1 since we don't know
-    # how many jobs such fleets can accommodate.
-    nodes_required_num = get_nodes_required_num_for_run(run_spec)
-    if (
-        fleet.spec.configuration.nodes is not None
-        and fleet.spec.configuration.blocks == 1
-        and fleet.spec.configuration.nodes.max is not None
-    ):
-        busy_instances = [i for i in fleet.instances if i.busy_blocks > 0]
-        fleet_available_capacity = fleet.spec.configuration.nodes.max - len(busy_instances)
-        if fleet_available_capacity < nodes_required_num:
-            return False
-    elif fleet.spec.configuration.ssh_config is not None:
-        # Currently assume that each idle block can run a job.
-        # TODO: Take resources / eligible offers into account.
-        total_idle_blocks = 0
-        for instance in fleet.instances:
-            total_blocks = instance.total_blocks or 1
-            total_idle_blocks += total_blocks - instance.busy_blocks
-        if total_idle_blocks < nodes_required_num:
-            return False
-    return True
-
-
-def _get_fleet_instance_offers(
-    fleet_model: FleetModel,
-    run_spec: RunSpec,
-    job: Job,
-    master_job_provisioning_data: Optional[JobProvisioningData] = None,
-    volumes: Optional[List[List[Volume]]] = None,
-) -> list[tuple[InstanceModel, InstanceOfferWithAvailability]]:
-    pool_instances = fleet_model.instances
-    instances_with_offers: list[tuple[InstanceModel, InstanceOfferWithAvailability]]
-    profile = run_spec.merged_profile
-    multinode = is_multinode_job(job)
-    nonshared_instances = filter_pool_instances(
-        pool_instances=pool_instances,
-        profile=profile,
-        requirements=job.job_spec.requirements,
-        status=InstanceStatus.IDLE,
-        fleet_model=fleet_model,
-        multinode=multinode,
-        master_job_provisioning_data=master_job_provisioning_data,
-        volumes=volumes,
-        shared=False,
-    )
-    instances_with_offers = [
-        (instance, common_utils.get_or_error(get_instance_offer(instance)))
-        for instance in nonshared_instances
-    ]
-    shared_instances_with_offers = get_shared_pool_instances_with_offers(
-        pool_instances=pool_instances,
-        profile=profile,
-        requirements=job.job_spec.requirements,
-        idle_only=True,
-        fleet_model=fleet_model,
-        multinode=multinode,
-        volumes=volumes,
-    )
-    instances_with_offers.extend(shared_instances_with_offers)
-    instances_with_offers.sort(key=lambda instance_with_offer: instance_with_offer[0].price or 0)
-    return instances_with_offers
 
 
 async def _assign_job_to_fleet_instance(
@@ -984,8 +754,8 @@ async def _run_jobs_on_new_instances(
     if fleet_model is not None:
         fleet = fleet_model_to_fleet(fleet_model)
         try:
-            _check_can_create_new_instance_in_fleet(fleet)
-            profile, requirements = _get_run_profile_and_requirements_in_fleet(
+            check_can_create_new_cloud_instance_in_fleet(fleet)
+            profile, requirements = get_run_profile_and_requirements_in_fleet(
                 job=job,
                 run_spec=run.run_spec,
                 fleet=fleet,
@@ -1112,43 +882,6 @@ async def _run_jobs_on_new_instances(
     return None
 
 
-def _get_run_profile_and_requirements_in_fleet(
-    job: Job,
-    run_spec: RunSpec,
-    fleet: Fleet,
-) -> tuple[Profile, Requirements]:
-    profile = combine_fleet_and_run_profiles(fleet.spec.merged_profile, run_spec.merged_profile)
-    if profile is None:
-        raise ValueError("Cannot combine fleet profile")
-    fleet_requirements = get_fleet_requirements(fleet.spec)
-    requirements = combine_fleet_and_run_requirements(
-        fleet_requirements, job.job_spec.requirements
-    )
-    if requirements is None:
-        raise ValueError("Cannot combine fleet requirements")
-    return profile, requirements
-
-
-def _check_can_create_new_instance_in_fleet(fleet: Fleet):
-    if not _can_create_new_instance_in_fleet(fleet):
-        raise ValueError("Cannot fit new instance into fleet")
-
-
-def _can_create_new_instance_in_fleet(fleet: Fleet) -> bool:
-    if fleet.spec.configuration.ssh_config is not None:
-        return False
-    active_instances = [i for i in fleet.instances if i.status.is_active()]
-    # nodes.max is a soft limit that can be exceeded when provisioning concurrently.
-    # The fleet consolidation logic will remove redundant nodes eventually.
-    if (
-        fleet.spec.configuration.nodes is not None
-        and fleet.spec.configuration.nodes.max is not None
-        and len(active_instances) >= fleet.spec.configuration.nodes.max
-    ):
-        return False
-    return True
-
-
 async def _create_fleet_model_for_job(
     exit_stack: AsyncExitStack,
     session: AsyncSession,
@@ -1158,7 +891,7 @@ async def _create_fleet_model_for_job(
     placement = InstanceGroupPlacement.ANY
     if run.run_spec.configuration.type == "task" and run.run_spec.configuration.nodes > 1:
         placement = InstanceGroupPlacement.CLUSTER
-    nodes = get_nodes_required_num_for_run(run.run_spec)
+    nodes = get_nodes_required_num(run.run_spec)
     lock_namespace = f"fleet_names_{project.name}"
     if is_db_sqlite():
         # Start new transaction to see committed changes after lock
